Modulerize BACKEND
* Changed most of API definition to a module
